Added tag jdk-22+10 for changeset 242a2e63

+++ src/hotspot/share/prims/jvm.cpp
@@ -608,7 +608,7 @@ JVM_END
 
 
 JVM_ENTRY(jint, JVM_IHashCode(JNIEnv* env, jobject handle))
-  // as implemented in the classic virtual machine; return 0 if object is nullptr
+  // as implemented in the classic virtual machine; return 0 if object is null
   if (handle == nullptr) {
     return 0;
   }
@@ -629,7 +629,7 @@ JVM_ENTRY(jint, JVM_IHashCode(JNIEnv* env, jobject handle))
       }
       return result.get_jint();
   } else {
-    return ObjectSynchronizer::FastHashCode(THREAD, obj);
+    return checked_cast<jint>(ObjectSynchronizer::FastHashCode(THREAD, obj));
   }
 JVM_END
 
@@ -1630,7 +1630,7 @@ static bool jvm_get_field_common(jobject field, fieldDescriptor& fd) {
   int modifiers = java_lang_reflect_Field::modifiers(reflected);
 
   InstanceKlass* ik = InstanceKlass::cast(k);
-  intptr_t offset = ik->field_offset(slot);
+  int offset = ik->field_offset(slot);
 
   if (modifiers & JVM_ACC_STATIC) {
     // for static fields we only look in the current class
